updated bootstrapper

+++ bootstrap.sh
@@ -1,20 +1,16 @@
 #!/bin/bash
 
 echo "symlink .bash_profile"
-rm -f ~/.bash_profile
-ln -s ~/.dotfiles/.bash_profile ~/.bash_profile
+[ ! -f ~/.bash_profile ] && ln -s ~/.dotfiles/.bash_profile ~/.bash_profile
 
 echo "symlink .screenrc"
-rm -f ~/.screenrc
-ln -s ~/.dotfiles/.screenrc ~/.screenrc
+[ ! -f ~/.screenrc ] && ln -s ~/.dotfiles/.screenrc ~/.screenrc
 
 echo "symlink .gitconfig"
-rm -f ~/.gitconfig
-ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
+[ ! -f ~/.gitconfig ] && ln -s ~/.dotfiles/.gitconfig ~/.gitconfig
 
 echo "symlink .gitignore"
-rm -f ~/.gitignore
-ln -s ~/.dotfiles/.gitignore ~/.gitignore
+[ ! -f ~/.gitignore ] && ln -s ~/.dotfiles/.gitignore ~/.gitignore
 
 echo "symlink Sublime Text preferences and keymap"
 rm -f ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User/Preferences.sublime-settings
@@ -22,9 +18,8 @@ ln -s ~/.dotfiles/Preferences.sublime-settings ~/Library/Application\ Support/Su
 rm -f ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User/Default\ \(OSX\).sublime-keymap
 ln -s ~/.dotfiles/Default\ \(OSX\).sublime-keymap ~/Library/Application\ Support/Sublime\ Text\ 2/Packages/User/Default\ \(OSX\).sublime-keymap
 
-echo "symlink Sublime Text command line"
-rm -f /usr/local/bin/subl
-ln -s /Applications/Sublime\ Text\ 2.app/Contents/SharedSupport/bin/subl /usr/local/bin/subl
+echo "symlink Sublime Text 2 command line"
+[ ! -f /usr/local/bin/subl ] && ln -s /Applications/Sublime\ Text\ 2.app/Contents/SharedSupport/bin/subl /usr/local/bin/subl
 
-echo "setup osx settings"
+echo "setup os x settings"
 ./.osx
